refactor: Folder structure
- Moved all components to components folder, keeping separate from the pages
