+++ src/components/SkillCard.tsx
