+++ src/pages/home/page.tsx
@@ -1,16 +1,18 @@
 import React from "react";
-import Hero from "@/pages/home/components/hero/Hero";
-import Skills from "@/pages/home/components/skills/Skills";
-import Technologies from "@/pages/home/components/technologies/Technologies";
-import Projects from "@/pages/home/components/projects/Projects";
-import Experience from "@/pages/home/components/experience/Experience";
-import Testimonials from "@/pages/home/components/testimonials/Testimonials";
-import About from "@/pages/home/components/about/About";
-import Footer from "@/pages/home/components/footer/Footer";
+import Navigation from "@/components/Navigation";
+import Hero from "@/components/Hero";
+import Skills from "@/components/Skills";
+import Technologies from "@/components/Technologies";
+import Projects from "@/components/Projects";
+import Experience from "@/components/Experience";
+import Testimonials from "@/components/Testimonials";
+import About from "@/components/About";
+import Footer from "@/components/Footer";
 
 const HomePage: React.FC = () => {
   return (
     <div className="w-full flex flex-col gap-20">
+      <Navigation />
       <Hero />
       <Skills />
       <Technologies />
